backend/requirements.txt
fastapi
uvicorn
google-generativeai
gTTS
langdetect